Fix Build And Update To Production Safe SDK and API (#1)
Currently running `make host` on a clean checkout doesn't work. 

The reason being that `index.html` is copied in the beginning of the makefile (at `npm run bundle`) and therefore the `index.html` target (that does the crucial replacements of style and js code) believes it has nothing more to do since `dist/index.html` already exists at this point. This leads to the Safe App loading with a blank screen (as no logic is injected).

Additionally, this PR modifies the app to use the latest SDK version and point to the production Safe transaction services (it used to point to the EIP-1271 preview services while they were under development).